requirements.txt
Flask
Flask-Login
Flask-SQLAlchemy
Werkzeug
Flask-Migrate
Flask-Bcrypt
Flask-SocketIO
python-dotenv
gunicorn
gevent
gevent-websocket
